# backend/docs/swagger.yaml
basePath: /
definitions:
...
  response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      success:
        type: boolean
    type: object
  response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  workflow.commentBody:
    properties:
      comment:
        type: string
    type: object
info:
  contact: {}
  description: K3 incident reporting and compliance API (MVP).
  title: Aegis API
  version: "1.0"
paths:
...
  /incidents/{id}/close:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Close incident
      tags:
      - workflow
...
  /incidents/{id}/reject:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: comment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/workflow.commentBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Reject incident
      tags:
      - workflow
  /incidents/{id}/start-corrective-action:
    post:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Move incident to corrective action (P2 stub)
      tags:
      - workflow
...
  /incidents/{id}/timeline:
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Incident workflow timeline
      tags:
      - workflow
  /incidents/{id}/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: comment
        in: body
        name: request
        schema:
          $ref: '#/definitions/workflow.commentBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Verify incident (supervisor)
      tags:
      - workflow
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"